+++ lib/app/pages/main/widgets/about_us_page.dart
@@ -1,11 +1,9 @@
 import 'package:animate_do/animate_do.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:merkar/app/core/resources/app_colors.dart';
-import 'package:merkar/app/core/resources/app_styles.dart';
 import 'package:merkar/app/core/resources/constants.dart';
 import 'package:merkar/app/core/resources/strings.dart';
-import 'package:merkar/app/core/widgets/internal_button.dart';
+import 'package:merkar/app/widgets/primary_button.dart';
 import 'package:url_launcher/url_launcher.dart';
 
 Future<void> AboutUsPage(BuildContext context) {
@@ -126,7 +124,7 @@ Future<void> AboutUsPage(BuildContext context) {
                                 ),
                               ],
                             ),
-                            InternalButton(title: Strings.label_close,
+                            PrimaryButton(title: Strings.label_close,
                               onPressed: () {
                                 Navigator.pop(context);
                               }),
