+++ lib/app/pages/main/widgets/comment_page.dart
@@ -1,11 +1,9 @@
 import 'package:animate_do/animate_do.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_email_sender/flutter_email_sender.dart';
-import 'package:merkar/app/core/resources/app_colors.dart';
-import 'package:merkar/app/core/resources/app_styles.dart';
 import 'package:merkar/app/core/resources/constants.dart';
 import 'package:merkar/app/core/resources/strings.dart';
-import 'package:merkar/app/core/widgets/internal_button.dart';
+import 'package:merkar/app/widgets/primary_button.dart';
 
 String? _message = "";
 Future<void> CommentePage(BuildContext context) {
@@ -50,7 +48,7 @@ Future<void> CommentePage(BuildContext context) {
                                     return null;
                                 },
                               ),
-                              InternalButton(title: Strings.label_send,
+                              PrimaryButton(title: Strings.label_send,
                                 onPressed: () {
                                 if (keyFormComments.currentState!
                                     .validate()) {
